Boring warning fixes.

+++ server/TracyFilesystem.hpp
@@ -17,7 +17,7 @@ static inline bool SourceFileValid( const char* fn, uint64_t olderThan )
     struct stat buf;
     if( stat( fn, &buf ) == 0 && ( buf.st_mode & S_IFREG ) != 0 )
     {
-        return buf.st_mtime < olderThan;
+        return (uint64_t)buf.st_mtime < olderThan;
     }
     return false;
 }
